fixed bug: colors float->int

+++ src/momapy/coloring.py
@@ -3,10 +3,10 @@ from dataclasses import dataclass, replace
 
 @dataclass(frozen=True)
 class Color(object):
-    red: float
-    green: float
-    blue: float
-    alpha: float = 1.0
+    red: int
+    green: int
+    blue: int
+    alpha: int = 1.0
 
     def __or__(self, alpha: float) -> "Color":
         if alpha < 0 or alpha > 100:
